requirements.txt
asgiref==3.8.1
asn1crypto==1.2.0
boto3==1.34.107
botocore==1.34.107
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap5==2024.2
cryptography==42.0.7
Django==5.0.6
django-crispy-forms==2.1
django-environ==0.11.2
django-storages==1.14.3
django-taggit==5.0.1
djangorestframework==3.15.1
docutils==0.15.2
gunicorn==22.0.0
idna==2.8
install==1.3.5
jmespath==0.10.0
packaging==24.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.1.0
python-dateutil==2.9.0.post0
requests==2.31.0
s3transfer==0.10.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==1.25.7
